refactor the project file structure
