+++ src/common/routes/Routes.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { Navigate, Route, Routes } from 'react-router-dom';
-import { Login } from '../../features/login/Login';
-import { Registration } from '../../features/registration/Registration';
-import { PasswordRecovery } from '../../features/password-recovery/PasswordRecovery';
+import { Login } from '../../features/auth/login/Login';
+import { Registration } from '../../features/auth/registration/Registration';
+import { PasswordRecovery } from '../../features/auth/password-recovery/PasswordRecovery';
 import { Profile } from '../../features/profile/Profile';
 import { ErrorPage } from '../404-page/ErrorPage';
-import { NewPassword } from '../../features/new-password/NewPassword';
+import { NewPassword } from '../../features/auth/new-password/NewPassword';
 import { Packs } from '../../features/Packs/Packs';
 import { Cards } from '../../features/Cards/Cards';
