+++ src/features/auth/login/Login.tsx
@@ -4,8 +4,8 @@ import FormControlLabel from '@mui/material/FormControlLabel';
 import Button from '@mui/material/Button';
 import { useFormik } from 'formik';
 import styles from './Login.module.css';
-import { loginTC } from './login-reducer';
-import { useAppDispatch, useAppSelector } from '../../app/store';
+import { loginTC } from '../auth-reducer';
+import { useAppDispatch, useAppSelector } from '../../../app/store';
 import { Link, Navigate } from 'react-router-dom';
 import { Visibility, VisibilityOff } from '@mui/icons-material';
 import { FormControl, IconButton, Input, InputAdornment, InputLabel } from '@mui/material';
