--- src/features/password-recovery/passwordRecovery-reducer.ts
@@ -1,21 +1,0 @@
-const initialState = {};
-
-export type InitialStateType = typeof initialState
-
-export const passwordRecoveryReducer = (
-  state: InitialStateType = initialState, action: PasswordRecoveryActionType): InitialStateType => {
-  switch (action.type) {
-    case 'SET-NEW-PASSWORD':
-      return { ...state };
-    default:
-      return state;
-  }
-};
-
-export type PasswordRecoveryActionType =
-  | SetNewPasswordACType
-
-export type SetNewPasswordACType = ReturnType<typeof setNewPasswordAC>
-
-export const setNewPasswordAC = () => ({ type: 'SET-NEW-PASSWORD' } as const);
-
